requirements.txt
streamlit
nltk
groq
pymongo
python-dotenv
pandas
scikit-learn
pyautogen>=0.2.0
